requirements.txt
# core
numpy==1.26.4
opencv-python==4.8.1.78
scipy==1.12.0
Pillow==10.2.0

# hardware interfaces
RPi.GPIO==0.7.1; sys_platform == "linux"
smbus2==0.4.3
pyserial==3.5
adafruit-circuitpython-pca9685==3.4.10

# database / storage
SQLAlchemy==2.0.28
flatbuffers==23.5.26

# vision / ML
face-recognition==1.3.0
onnxruntime==1.17.1

# audio
pyaudio==0.2.14
sounddevice==0.4.6

# utilities
pyyaml==6.0.1
click==8.1.7
structlog==24.1.0
watchdog==4.0.0

# testing
pytest==8.0.2
pytest-mock==3.12.0
pytest-asyncio==0.23.5
